@@ -1,6 +1,6 @@
-/// Get Instance Manager is a modern and intelligent dependency injector
+/// JetX Instance Manager is a modern and intelligent dependency injector
 /// that injects and removes dependencies seasonally.
 library;
 
-export 'get_core/get_core.dart';
-export 'get_instance/get_instance.dart';
+export 'jetx_core/jetx_core.dart';
+export 'jetx_instance/jetx_instance.dart';
