@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:get/get.dart';
+import 'package:jetx/jetx.dart';
 
 class Wrapper extends StatelessWidget {
   final Widget? child;
-  final List<GetPage>? namedRoutes;
+  final List<JetPage>? namedRoutes;
   final String? initialRoute;
   final Transition? defaultTransition;
 
@@ -17,7 +17,7 @@ class Wrapper extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return GetMaterialApp(
+    return JetMaterialApp(
       defaultTransition: defaultTransition,
       initialRoute: initialRoute,
       translations: WrapperTranslations(),
@@ -34,7 +34,7 @@ class Wrapper extends StatelessWidget {
 
 class WrapperNamed extends StatelessWidget {
   final Widget? child;
-  final List<GetPage>? namedRoutes;
+  final List<JetPage>? namedRoutes;
   final String? initialRoute;
   final Transition? defaultTransition;
 
@@ -48,7 +48,7 @@ class WrapperNamed extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return GetMaterialApp(
+    return JetMaterialApp(
       defaultTransition: defaultTransition,
       initialRoute: initialRoute,
       getPages: namedRoutes,
