@@ -1,6 +1,6 @@
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:get/utils.dart';
+import 'package:jetx/utils.dart';
 
 class Foo extends StatelessWidget {
   const Foo({super.key});
